image/pyproject.toml
    [project]
    name = "mcp-image-tools"
    version = "1.0.0"
    description = "Image processing tools for MCP — read, resize, crop, convert, return as base64"
    readme = "README.md"
    license = "MIT"
    requires-python = ">=3.10"
    authors = [
        { name = "Ulrich Cuber", email = "sonstige@uc-it.de" },
    ]
    keywords = ["mcp", "image", "tools"]
    classifiers = [
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3",
    ]
    dependencies = [
        "mcp-server-framework>=1.3.0",
"Pillow>=10.0.0",
    ]

    [project.urls]
    Homepage = "https://github.com/cuber-it/mcp_tools"
    Repository = "https://github.com/cuber-it/mcp_tools"

    [project.scripts]
    mcp-image-tools = "mcp_image_tools:main"

    [build-system]
    requires = ["hatchling"]
    build-backend = "hatchling.build"

    [tool.hatch.build.targets.wheel]
    packages = ["src/mcp_image_tools"]
